+++ crates/ibc/Cargo.toml
@@ -14,18 +14,26 @@ version.workspace = true
 
 [features]
 default = []
+migrations = [
+  "namada_migrations",
+  "linkme",
+]
 testing = ["namada_core/testing", "ibc-testkit", "proptest"]
 
 [dependencies]
 namada_core = { path = "../core" }
 namada_gas = { path = "../gas" }
 namada_governance = { path = "../governance" }
+namada_macros = {path = "../macros"}
+namada_migrations = {path = "../migrations", optional = true}
 namada_parameters = { path = "../parameters" }
 namada_state = { path = "../state" }
 namada_storage = { path = "../storage" }
 namada_token = { path = "../token" }
 
 borsh.workspace = true
+konst.workspace = true
+linkme = {workspace = true, optional = true}
 ibc.workspace = true
 ibc-derive.workspace = true
 ibc-testkit = {workspace = true, optional = true}
@@ -34,6 +42,7 @@ masp_primitives.workspace = true
 primitive-types.workspace = true
 proptest = { workspace = true, optional = true }
 prost.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
