+++ crates/ibc/src/actions.rs
@@ -11,7 +11,7 @@ use namada_core::ibc::apps::transfer::types::msgs::transfer::MsgTransfer as IbcM
 use namada_core::ibc::apps::transfer::types::packet::PacketData;
 use namada_core::ibc::apps::transfer::types::PrefixedCoin;
 use namada_core::ibc::core::channel::types::timeout::TimeoutHeight;
-use namada_core::ibc::{IbcEvent, MsgTransfer};
+use namada_core::ibc::MsgTransfer;
 use namada_core::tendermint::Time as TmTime;
 use namada_core::token::Amount;
 use namada_governance::storage::proposal::PGFIbcTarget;
@@ -23,6 +23,7 @@ use namada_state::{
 use namada_token as token;
 use token::DenominatedAmount;
 
+use crate::event::IbcEvent;
 use crate::{IbcActions, IbcCommonContext, IbcStorageContext};
 
 /// IBC protocol context
