+++ crates/tx_prelude/src/ibc.rs
@@ -6,8 +6,8 @@ use std::rc::Rc;
 
 use namada_core::address::{Address, InternalAddress};
 use namada_core::event::EventTypeBuilder;
-pub use namada_core::ibc::IbcEvent;
 use namada_core::token::Amount;
+pub use namada_ibc::event::{IbcEvent, IbcEventType};
 pub use namada_ibc::storage::{ibc_token, is_ibc_key};
 pub use namada_ibc::{
     IbcActions, IbcCommonContext, IbcStorageContext, NftTransferModule,
